requirements.txt
pytorch>=1.0.0
torchvision>=0.2.1
python=3.6
fenics=2018.1.0
matplotlib
h5py
json